[Breaking] Rename confusing terms such as `Split` (#67)
This PR renames `Split` to `SubClause` and `TreePath` to `Clause`. In
turn, this renaming makes things much more intuitive since it sticks to
the end-result terms instead of internal representation (decision tree)
terms. Also, the `Split` datastructure is finally fixed to directly
contain the split value and feature instead of having those in a nested
data structure called `SplitPoint`.